+++ src/components/text-editor/prosemirror-adapter/menu/menu-commands.ts
@@ -1,6 +1,109 @@
 import { toggleMark, setBlockType, wrapIn } from 'prosemirror-commands';
-import { Schema, MarkType } from 'prosemirror-model';
+import { Schema, MarkType, NodeType } from 'prosemirror-model';
 import { wrapInList } from 'prosemirror-schema-list';
+import { Command } from 'prosemirror-state';
+import { EditorMenuTypes, LevelMapping } from './types';
+
+type CommandFunction = (
+    schema: Schema,
+    mark: EditorMenuTypes,
+    url?: string,
+) => Command;
+
+interface CommandMapping {
+    [key: string]: CommandFunction;
+}
+
+const createToggleMarkCommand = (
+    schema: Schema,
+    markName: string,
+    url?: string,
+): Command => {
+    const markType: MarkType | undefined = schema.marks[markName];
+    if (!markType) {
+        throw new Error(`Mark "${markName}" not found in schema`);
+    }
+
+    if (markName === EditorMenuTypes.Link && url) {
+        const attrs = {
+            href: url,
+            target: isExternalLink(url) ? '_blank' : null,
+        };
+
+        return toggleMark(markType, attrs);
+    }
+
+    return toggleMark(markType);
+};
+
+const isExternalLink = (url: string): boolean => {
+    return !url.startsWith(window.location.origin);
+};
+
+const createSetNodeTypeCommand = (
+    schema: Schema,
+    nodeType: string,
+    level?: number,
+): Command => {
+    const type: NodeType | undefined = schema.nodes[nodeType];
+    if (!type) {
+        throw new Error(`Node type "${nodeType}" not found in schema`);
+    }
+
+    if (nodeType === 'heading' && level) {
+        return setBlockType(type, { level: level });
+    } else {
+        return setBlockType(type);
+    }
+};
+
+const createWrapInCommand = (schema: Schema, nodeType: string): Command => {
+    const type: NodeType | undefined = schema.nodes[nodeType];
+    if (!type) {
+        throw new Error(`Node type "${nodeType}" not found in schema`);
+    }
+
+    return wrapIn(type);
+};
+
+const createListCommand = (schema: Schema, listType: string): Command => {
+    const type: NodeType | undefined = schema.nodes[listType];
+    if (!type) {
+        throw new Error(`List type "${listType}" not found in schema`);
+    }
+
+    return wrapInList(type);
+};
+
+const commandMapping: CommandMapping = {
+    strong: createToggleMarkCommand,
+    em: createToggleMarkCommand,
+    underline: createToggleMarkCommand,
+    headerlevel1: (schema) =>
+        createSetNodeTypeCommand(
+            schema,
+            LevelMapping.Heading,
+            LevelMapping.one,
+        ),
+    headerlevel2: (schema) =>
+        createSetNodeTypeCommand(
+            schema,
+            LevelMapping.Heading,
+            LevelMapping.two,
+        ),
+    headerlevel3: (schema) =>
+        createSetNodeTypeCommand(
+            schema,
+            LevelMapping.Heading,
+            LevelMapping.three,
+        ),
+    blockquote: createWrapInCommand,
+    /* eslint-disable camelcase */
+    ordered_list: createListCommand,
+    bullet_list: createListCommand,
+    /* eslint-enable camelcase */
+    link: createToggleMarkCommand,
+};
 
 export class MenuCommandFactory {
     private schema: Schema;
@@ -9,85 +112,22 @@ export class MenuCommandFactory {
         this.schema = schema;
     }
 
-    private markNames = {
-        bold: 'strong',
-        italic: 'em',
-        underline: 'underline',
-        blockquote: 'blockquote',
-        headerlevel1: 'headerlevel1',
-        headerlevel2: 'headerlevel2',
-        headerlevel3: 'headerlevel3',
-        addorremovelink: 'link',
-        numberedlist: 'ordered_list',
-        list: 'bullet_list',
-    };
-
-    createCommand(mark: string) {
-        if (this.markNames[mark]) {
-            mark = this.markNames[mark];
+    public getCommand(mark: EditorMenuTypes, url?: string) {
+        const commandFunc = commandMapping[mark];
+        if (!commandFunc) {
+            throw new Error(`The Mark "${mark}" is not supported`);
         }
 
-        switch (mark) {
-            case 'strong':
-            case 'em':
-            case 'underline':
-                return this.createToggleMarkCommand(mark);
-            case 'paragraph':
-                return this.createSetNodeTypeCommand(mark);
-            case 'headerlevel1':
-            case 'headerlevel2':
-            case 'headerlevel3':
-                return this.createSetNodeTypeCommand(
-                    'heading',
-                    parseInt(mark[mark.length - 1], 10),
-                );
-            case 'blockquote':
-                return this.createWrapInCommand(mark);
-            case 'ordered_list':
-            case 'bullet_list':
-                return this.createListCommand(mark);
-            default:
-                throw new Error(`The Mark "${mark}" is not supported`);
-        }
+        return commandFunc(this.schema, mark, url);
     }
 
-    private createToggleMarkCommand(markName: string) {
-        const markType: MarkType | undefined = this.schema.marks[markName];
-        if (!markType) {
-            throw new Error(`Mark "${markName}" not found in schema`);
-        }
-
-        return toggleMark(markType);
-    }
-
-    private createSetNodeTypeCommand(nodeType: string, level?: number) {
-        const type = this.schema.nodes[nodeType];
-        if (!type) {
-            throw new Error(`Node type "${nodeType}" not found in schema`);
-        }
-
-        if (nodeType === 'heading' && level) {
-            return setBlockType(type, { level: level });
-        } else {
-            return setBlockType(type);
-        }
-    }
-
-    private createWrapInCommand(nodeType: string) {
-        const type = this.schema.nodes[nodeType];
-        if (!type) {
-            throw new Error(`Node type "${nodeType}" not found in schema`);
-        }
-
-        return wrapIn(type);
-    }
-
-    private createListCommand(listType: string) {
-        const type = this.schema.nodes[listType];
-        if (!type) {
-            throw new Error(`List type "${listType}" not found in schema`);
-        }
-
-        return wrapInList(type);
+    buildKeymap() {
+        return {
+            'Mod-B': this.getCommand(EditorMenuTypes.Bold),
+            'Mod-I': this.getCommand(EditorMenuTypes.Italic),
+            'Mod-Shift-1': this.getCommand(EditorMenuTypes.HeaderLevel1),
+            'Mod-Shift-2': this.getCommand(EditorMenuTypes.HeaderLevel2),
+            'Mod-Shift-3': this.getCommand(EditorMenuTypes.HeaderLevel3),
+        };
     }
 }